refactor: File for each browser setup, adds tests
