+++ packages/wdio-utils/src/driver/edge.ts
@@ -1,0 +1,6 @@
+import { download as downloadEdgedriver } from 'edgedriver'
+export type { EdgedriverParameters } from 'edgedriver'
+
+export function setupEdgedriver(cacheDir: string, driverVersion?: string) {
+    return downloadEdgedriver(driverVersion, cacheDir)
+}
